Fixed MSVC 6.0 warnings

+++ sf-cvs/trunk/audacity-src/src/effects/Phaser.cpp
@@ -100,7 +100,7 @@ bool EffectPhaser::NewTrackSimpleMono()
 
    phase = startphase;
    if (mCurChannel == Track::RightChannel)
-      phase += M_PI;
+      phase += (float)M_PI;
 
    return true;
 }
